+++ src/main/java/com/sber/practika/models/requests/authentication/AuthenticationRequestBankCard.java
@@ -6,6 +6,6 @@ import java.math.BigInteger;
 
 @Data
 public class AuthenticationRequestBankCard {
-    private BigInteger bankCard;
+    private Long bankCard;
     private String password;
 }
