+++ src/main/java/com/sber/practika/models/requests/finalize/FinalizeTransferRequestBankCardBetweenBankNumber.java
@@ -2,13 +2,12 @@ package com.sber.practika.models.requests.finalize;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.UUID;
 
 @Data
 public class FinalizeTransferRequestBankCardBetweenBankNumber {
     private UUID uuidTransaction;
-    private BigInteger bankCard;
+    private Long bankCard;
     private String bankNumber;
-    private BigInteger value;
+    private Long value;
 }
