+++ src/main/java/com/sber/practika/models/requests/finalize/FinalizeTransferRequestBankCardToBankCard.java
@@ -2,13 +2,12 @@ package com.sber.practika.models.requests.finalize;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.UUID;
 
 @Data
 public class FinalizeTransferRequestBankCardToBankCard {
     private UUID uuidTransaction;
-    private BigInteger bankCard1;
-    private BigInteger bankCard2;
-    private BigInteger value;
+    private Long bankCard1;
+    private Long bankCard2;
+    private Long value;
 }
