+++ src/main/java/com/sber/practika/models/requests/finalize/FinalizeTransferRequestBankNumberToBankNumber.java
@@ -2,7 +2,6 @@ package com.sber.practika.models.requests.finalize;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.util.UUID;
 
 @Data
@@ -10,5 +9,5 @@ public class FinalizeTransferRequestBankNumberToBankNumber {
     private UUID uuidTransaction;
     private String bankNumber1;
     private String bankNumber2;
-    private BigInteger value;
+    private Long value;
 }
