+++ src/main/java/com/sber/practika/models/requests/support/confirmCode/SupportRequestConfirmCodeAndBankCard.java
@@ -2,9 +2,8 @@ package com.sber.practika.models.requests.support.confirmCode;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 @Data
 public class SupportRequestConfirmCodeAndBankCard {
-    private BigInteger bankCard;
+    private Long bankCard;
     private Integer confirmCode;
 }
