+++ src/main/java/com/sber/practika/models/requests/support/confirmCode/SupportRequestConfirmCodeAndPhone.java
@@ -2,9 +2,8 @@ package com.sber.practika.models.requests.support.confirmCode;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 @Data
 public class SupportRequestConfirmCodeAndPhone {
-    private BigInteger phone;
+    private Long phone;
     private Integer confirmCode;
 }
