+++ src/main/java/com/sber/practika/models/requests/support/newPassword/SupportRequestNewPasswordAndBankCard.java
@@ -2,10 +2,8 @@ package com.sber.practika.models.requests.support.newPassword;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class SupportRequestNewPasswordAndBankCard {
-    private BigInteger bankCard;
+    private Long bankCard;
     private String newPassword;
 }
