+++ src/main/java/com/sber/practika/models/requests/support/newPassword/SupportRequestNewPasswordAndPhone.java
@@ -2,10 +2,8 @@ package com.sber.practika.models.requests.support.newPassword;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class SupportRequestNewPasswordAndPhone {
-    private BigInteger phone;
+    private Long phone;
     private String newPassword;
 }
