+++ src/main/java/com/sber/practika/models/requests/transfer/TransferRequestBankCardBetweenBankNumber.java
@@ -2,11 +2,9 @@ package com.sber.practika.models.requests.transfer;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class TransferRequestBankCardBetweenBankNumber {
-    private BigInteger bankCard;
+    private Long bankCard;
     private String bankNumber;
-    private BigInteger value;
+    private Long value;
 }
