+++ src/main/java/com/sber/practika/models/requests/transfer/TransferRequestBankCardBetweenPhone.java
@@ -2,11 +2,9 @@ package com.sber.practika.models.requests.transfer;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class TransferRequestBankCardBetweenPhone {
-    private BigInteger bankCard;
-    private BigInteger phone;
-    private BigInteger value;
+    private Long bankCard;
+    private Long phone;
+    private Long value;
 }
