+++ src/main/java/com/sber/practika/models/requests/transfer/TransferRequestBankNumberBetweenBankNumber.java
@@ -2,11 +2,9 @@ package com.sber.practika.models.requests.transfer;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class TransferRequestBankNumberBetweenBankNumber {
     private String bankNumber1;
     private String bankNumber2;
-    private BigInteger value;
+    private Long value;
 }
