+++ src/main/java/com/sber/practika/models/requests/transfer/TransferRequestBankNumberBetweenPhone.java
@@ -2,11 +2,9 @@ package com.sber.practika.models.requests.transfer;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class TransferRequestBankNumberBetweenPhone {
     private String bankNumber;
-    private BigInteger phone;
-    private BigInteger value;
+    private Long phone;
+    private Long value;
 }
