+++ src/main/java/com/sber/practika/models/requests/transfer/TransferRequestPhoneBetweenPhone.java
@@ -2,11 +2,9 @@ package com.sber.practika.models.requests.transfer;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class TransferRequestPhoneBetweenPhone {
-    private BigInteger phone1;
-    private BigInteger phone2;
-    private BigInteger value;
+    private Long phone1;
+    private Long phone2;
+    private Long value;
 }
